Only output details of commits that HE strings
High-entropy strings that appear in the "context" of diffs, or those
that are removed by a commit aren't particularly interesting when
you're searching a repository's history for secrets. This change means
trufflehog only searches diffs for lines that introduce potential
secrets, dramatically reducing the amount of noise to search through
in trufflehog's output.