[Fix] Upgraded AWS go-sdk to V2 (only for AWS Canary Token Verification) (#3907)
* upgrade go-aws-sdk to v2 as the v1 is about to be deprecated and also returning errors randomly when verify canary tokens.
ran go mod tidy.

* accidently removed the old version. readding old version

* use aws go sdk v2 for access key verification

* remove unused param

* updated comments.

* use verification client for better logging.
SDK has changed the error if the provided Id is invalid, incorporated the handling for that error as well.

* resolved the http client issue using with aws sdk.
passing the user-agent via middleware in aws sdk.
remove the hash logic since it was making the AWS Access key test as flaky by randomly generating the incorrect values.

* re-added the retry logic if AWS returns 403