Include Context Lines In Git Diff Chunking
Previously we sort of ignored context lines when generating diff chunks to hand off
to the detectors. (We actually would add a new line for each? Not sure why). Now we
include a single context line (one before and after the diff itself) in the diff
chunk passed off to the detectors.

This is important because, in some cases, the match key (that is used to determine
which detectors to pass diff chunks to) isn't otherwise included in the diff chunk,
as when a commit updates a private key in a previously existing cert file.  The
header, `-----BEGIN RSA PRIVATE KEY-----`, isn't otherwise included in the diff.