Update Diff Chunk Tests To Be Context Line Aware
Updates the tests the touch diff chunk content to check for the new context
lines that we're adding to all diff chunks.  This change was a pain, and it's
a little ugly having these big long strings in our tests, but not sure it's
worth it/useful to try and do anything different.