Don't die on workspace request failure to Postman API
Right now when the Postman source is scanning the postman API, if it gets *any* error back from
the API it stops the whole scan.  We should probably just log and continue if this happens, so this
makes that change for requests to the *workspace* endpoint

This issue popped up when investigating a logged problem with making requests to the Postman API
and (likely) sending either a bad collection/workspace id or (also possible) sending a collection
or workspace id that we don't have permission to explore to the Postman API.  Still worth looking
more into that, but I think that making these errors less visible is worth the tradeoff for still
getting _something_ out of a scan.