Fix "skipping binary file" logging to show actual file name
This was intended to show path:fileName, but `path` in these scopes is
the repo path, not a file.  This section was moved in a refactor where
`path` was the file in the old scope, and since both scopes have `path
string`, it was not flagged and easy to miss.