Make Integrations team codeown source integrations (#4561)
We have a dedicated team for our source integrations. This commit modifies CODEOWNERS so that they own those integrations.

It does not fiddle with proto/ codeownership, because that's going to be a mess to untangle and isn't related to the actual review requests that motivated this change. It also doesn't make the Integrations team codeown our more experimental/unusual sources:

- GitHub Experimental
- Huggingface
- stdin
- syslog