Add more clone logging (#4593)
This commit adds timing information to the logs when clones complete successfully. There are several other more robust but also more involved interventions we could do, and I think we should do them, but this is a strict improvement that's easy to get out as a first step.

I also added logging at the precise beginning of the clone command so we know when we're doing that.